backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
langgraph==0.2.60
langchain-groq==0.2.4
langchain-core==0.3.28
pypdf==5.1.0
python-docx==1.1.2
chardet==5.2.0
tenacity==9.0.0
aiofiles==24.1.0
python-multipart==0.0.20
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.7.0
websockets==14.1
